fix uninitialized local storage (?) bookmark bug
Restoring datatable state would fail when loading the app for the first
time in a "fresh" browser, probably because state loading is skipped
when there's no state in local storage. However, if we load the state
data by overriding the stateLoadCallback(), instead of stateLoadParams()
(which just allows tweaks to be made to the state data after it's been
loaded in the standard way), we explicitly tell DataTables that the data
will be coming from elsewhere, so uninitialized local storage doesn't
matter.