implement model comparison feature
As part of this, data loading was extracted into its own function
`loadData()` in `load_data.R`, because for model comparison, we actually
need to load *two* result datasets. This will also make it easier to add
new functionality in the future, as the feature developer can easily
load the data for development and testing in the same format that the
application uses.