Accommodate CPACT results
CPACT results are expected to be in the same .RData format as koditex
results. It's basically the koditex results on which the CPACT results
are based, with CPACT rows added to the `factors` data frame. The `MODE`
of these rows should be `cpact`, their `DIVISION` should correspond to
the CPACT task.