Get rid of scrollable settings well
It's brittle, it interferes with drop-down lists, it just wasn't meant
to be. The problem it was trying to solve is that if the Factors 2-dim
tab isn't overflowing the viewport, there's no scrollbar; then if a
chunk is selected and it starts overflowing, a scrollbar appears, which
makes the viewport narrower, which makes the plot re-render, which
clears the selected chunk (bummer). A somewhat uglier but more reliable
solution to this problem is simply to always display a vertical
scrollbar on the body, even if the content is not overflowing the
viewport.